requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
redis==5.0.8
google-generativeai==0.8.3
tenacity==8.5.0
chromadb==0.5.5
sqlalchemy==2.0.35
PyPDF2==3.0.1
httpx==0.27.2
pydantic==2.5.0
jsonschema==4.23.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.21.0
numpy==1.26.4
python-multipart==0.0.9
gunicorn==21.2.0
